data-service/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic-settings>=2.4.0
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
python-multipart>=0.0.9
psycopg2-binary>=2.9.0
apscheduler>=3.10.0
